mistakes in the `Purify2to1Node` docstring
There was a bit of confusion here. The point of entanglement distillation is to purify pairs that are not all store on the same node. I.e. Alice has two halves of two different Bell pairs. Bob has the other halves. This is very different from Alice locally having a Bell pair and Bob locally having an independent Bell pair -- there is no shared entanglement in such a case and no possiblity for quantum networking.

A second issue is that the purification was reporting a failure in the docstring - this could have been a sign to investigate what the problem is.